requirements.txt
tensorflow==2.15.0
keras==2.15.0
streamlit==1.28.1
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
protobuf==6.31.1
